zoom-plugin-general/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
distro==1.9.0
docstring_parser==0.16
Flask==2.3.2
google==3.0.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.4
openai==1.25.0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==2.3.7
wheel==0.41.2
